Add only_if check to .bundle chown in dashboard recipe (#106)
__Why__

* Adding the subscribes attributes did not prevent the execute block from
executing in order.

__How__

* Add a only_if conditional to check to make sure that the .bundle dir exists
before changing the permissions.

+++ kitchen/cookbooks/login_dot_gov/recipes/dashboard.rb
@@ -20,7 +20,8 @@ sha_env = (node.chef_environment == 'dev' ? node['login_dot_gov']['branch_name']
 end
 
 execute "chown -R #{node['login_dot_gov']['system_user']}:nogroup #{base_dir}"
-execute "chown -R #{node['login_dot_gov']['system_user']} /home/ubuntu/.bundle" do
+execute "chown -R #{node['login_dot_gov']['system_user']} /home/#{node['login_dot_gov']['system_user']}/.bundle" do
+  only_if { ::Dir.exist?("/home/#{node['login_dot_gov']['system_user']}/.bundle") }
   subscribes :run, "execute[/opt/ruby_build/builds/#{node['login_dot_gov']['ruby_version']}/bin/bundle install --deployment --jobs 3 --path /srv/dashboard/shared/bundle --without deploy development test]", :immediately
   subscribes :run, "execute[/opt/ruby_build/builds/#{node['login_dot_gov']['ruby_version']}/bin/bundle install --deployment --jobs 3 --path /srv/sp-rails/shared/bundle --without deploy development test]", :immediately
   subscribes :run, "execute[/opt/ruby_build/builds/#{node['login_dot_gov']['ruby_version']}/bin/bundle install --deployment --jobs 3 --path /srv/sp-sinatra/shared/bundle --without deploy development test]", :immediately
